Moved legacy rake methods into utils folder
